requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl_cffi==0.7.1
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
google-ads==25.0.0
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
hikerapi==1.5.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_resources==6.4.5
instagrapi==2.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.1
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0